refactor: split gpt helper
