+++ src/helpers/onTextMessage.ts
@@ -7,7 +7,7 @@ import { log } from "../helpers.ts";
 import { addToHistory, forgetHistory } from "./history.ts";
 import { setLastCtx } from "./lastCtx.ts";
 import { addOauthToThread, ensureAuth } from "./google.ts";
-import { getChatgptAnswer } from "./gpt.ts";
+import { requestGptAnswer } from "./gpt.ts";
 import checkAccessLevel from "./checkAccessLevel.ts";
 import resolveChatButtons from "./resolveChatButtons.ts";
 import { sendTelegramMessage } from "./telegram.ts";
@@ -165,7 +165,7 @@ async function answerToMessage(
     let msgSent: Message.TextMessage | undefined;
     await ctx.persistentChatAction("typing", async () => {
       if (!msg) return;
-      const res = await getChatgptAnswer(msg, chat, ctx);
+      const res = await requestGptAnswer(msg, chat, ctx);
       const text = res?.content || "бот не ответил";
       const extraParams: Record<string, unknown> = {
         ...extraMessageParams,
