+++ tests/helpers/retryLogicSimple.test.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { jest } from "@jest/globals";
 
 describe("retryLogicSimple", () => {
